LU-18243 lfs: Add --output and --no-header options to lfs df command
Adds a new --output option to the lfs df command that allows
users to specify which fields to display and in what order.
This replaces the need for "lfs df | awk" patterns commonly
used in test scripts.

Also adds a --no-header (-N) option to suppress column headers
in the output, useful for scripting and automation where only
data values are needed.

Supported fields and aliases:
- source (device): filesystem UUID
- size (total, btotal): total space
- used (bused): used space
- free (bfree): free space
- avail (bavail): available space
- pcent (usepct, bpct, pct): usage percentage
- itotal (inodes): total inodes
- iused: used inodes
- iavail (ifree): available inodes
- ipcent (iusepct, ipct): inode usage percentage
- target: mount point

Headers are shown by default but can be suppressed using --no-header
or when using --output with a specific target index (e.g., -o0) for
easier scripting.

Examples:
  lfs df --output=avail,source         # shows headers
  lfs df --output=avail -o0            # no headers for scripting
  lfs df --output=avail --no-header    # no headers explicitly
  lfs df --no-header -o                # OST values without headers

Test cases updated to use the new --output and --no-header
functionality.

Signed-off-by: Chakshu Kansal <ckansal@ddn.com>
Change-Id: I659b5a11fbf1aaef179d7c47d1843b38ad22b832
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59700
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Olaf Faaland <faaland1@llnl.gov>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Frederick Dilger <fdilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>