LU-19237 llite: adopt i_version use to guard directory cache
NFSv4 (knfsd) uses i_version as a way to gauge if inode
have changed in addition to ctime, since we don't support
nanosecond timestamps, a lot of activity could go in that
othewise goes unnoticed.

So to combat stale readdir cache we are adding i_version use.

This is the simplest possible implementation where we let VFS
manage i_version on regular files for us and we only do it for
directories.
Furthermore we don't actually store i_version on disk as it's
not really important to our usecase I think, all we want to
ensure is that NFS exports are fine, zero use for our own benefit.
And besides transferring i_version from server is a protocol
change and othe hassles. See some more discussion in the ticket.

Change-Id: If8f678c98b5d2e7acf5b4c3928a6acbedbc52258
Test-Parameters: trivial
Signed-off-by: Oleg Drokin <green@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60563
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>