LU-19131 utils: 'lfs mirror extend' should use DIO
To reduce memory pressure on the data mover nodes, use O_DIRECT
when opening files.  Otherwise, this can cause issues on OSS
nodes or with highly concurrent data migraton.

Use the same buffer size for all mirror/migrate handling.
Fix lfs_mirror_verify_chunk() to handle a buffer size that
is not an even multiple of the component extent size.

Name the llapi_migration_flags enum and use it in the callers.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I085b64f93455c91e5a7498f36f815e9a6f300c1e
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60766
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Alexandre Ioffe <aioffe@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Rahul Bansal <rahulmay94@gmail.com>
Tested-by: jenkins <devops@whamcloud.com>