LU-19157 nodemap: revoke export locks on nodemap change
When a nodemap definition is changed, the locks for the member exports
should be revoked under certain conditions. Fix the current mechanism
in ldlm_revoke_export_locks() as it does nothing because -EOPNOTSUPP
is returned from cfs_hash_for_each_nolock().
Member exports associated with a nodemap should not be revoked after
changes occured on a different nodemap. And for member exports
associated with the nodemap being changed, not all changes require to
revoke locks. We care about clients caching permissions that are no
longer correct, so we are interested in nodemap properties such as
admin, trusted, squash ids, offsets, and id mappings as well.
And we also have to revoke locks for clients being banned.

sanity-sec test_81 is used to verify that the lock revoking process
works fine with banlist, and sanity-sec test_82 is added to exercise
the lock revoking process in case of nodemap changes.
Also fix sanity-sec test_21 to make sure lru locks are flushed after
nodemaps have been modified, and that this is done on all clients.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: Ibca0f280ea70dcd32256c4c6f5252cb6b6b18171
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60643
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@dilger.ca>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>