LU-19248 tests: Wait for fake intf ipv6 dup detection
IPv6 addresses undergo duplicate address detection during which they
are flagged as "tentative" in the output of ip addr show. Some test
cases do not work correctly until this flag is cleared on the fake
interface that is created in setup_fakeif(). Modify this function to
wait for the flag to be cleared before returning.

Test-Parameters: trivial
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Ie7217feb7ed3c407f66f3d4a48fbb954870dd521
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60677
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Frank Sehr <fsehr@whamcloud.com>
Reviewed-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>