LU-19270 llite: init cl before getting root inode on mount
In "client_common_fill_super()", "ll_iget()" retrieves the root inode
which accesses "cl_file_inode_init()" and "struct ll_sb_info->ll_site"
for regular files. "ll_site" is not initialized at that point causing
a panic when a file is used as the mounting path due to "Unable to
handle kernel NULL pointer dereference".

This patch fixes the panic by moving the "cl" initialization in
"client_common_fill_super()" before retrieving the root inode. Using a
file as the mount path can therefore show the error message that the
presented path is not a directory.

Further, this patch adds error handling for "cl_sb_init(sb)" and
finalizes it in the error path.

Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: Ie1f85173563cba6f13ca63ed4efe5f3f351b274c
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60801
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Patrick Farrell <pfarrell@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>