LU-19279 obdclass: add root fid to debugfs mdt exports
When clients are mounted, the fid of the mounted subdirectory (root
fid) is stored in the client export on the MDT0. This root_fid remains
set for the lifetime of the mounted client, but it is not currently
exposed.

This patch adds the root fid field from the client export to the "lctl
get_param mdt.TARGET.exports.NID.export" output if set. Therefore, it
is only listed for MDT0 and excluded for all other targets.

conf-sanity test_156 was added to test that the root_fid on the export
is consistent with the fid of the mounted directory on the client.

Test-Parameters: trivial testlist=sanity-sec,conf-sanity
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: I5a5a7a4d8404f24386b00a9ac601cb2e041108ee
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60844
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>