LU-19295 utils: handle insufficient arguments to lst
When passing insufficient arguments to lst command it
was expecting to throw a usage message. Instead it was
segfaulting. The reason was that cfs_parser() did not
have an exit condition and was recursively calling
the same function and finally reaching the stack limit
and segfaulting. This behavior was seen across all the
lst commands.

Add an exit condition and pass the return value to the
calling function so it can bail out with a usage message.

------------
Before patch
------------
$ lst add_group
Segmentation fault (core dumped)

$ lst del_group
Segmentation fault (core dumped)

------------
After patch
------------
$ lst add_group
Usage: lst group NAME IDs [IDs]...

$ lst del_group
Usage: lst del_group NAME

Test-Parameters: trivial
Fixes: 21080400f ("LU-16723 libcfs: refactor parser to be simpler")
Signed-off-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Change-Id: Iaedd068629f1f1699b5efad0104279eae9ef9907
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/60886
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>