LU-19066 obdfilter: add failure_domain parameter
Add a failure_domain parameter to the device.

This can be set on a per-device basis using:
  lctl set_param obdfilter.lustrewt-OST0000.failure_domain=55
and will be used to allow FLR stripe allocation
to become failure_domain aware.

Later patches will store this persistently on each OST itself
and will not be needed to be stored in the persistent MGS config
with "set_param -P".

Test-Parameters: trivial
Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>
Change-Id: Ib354f9351426e8af93e1ab813aee624c0eccb167
Was-Change-Id: I910b653d40f592daded28b4063f19aedc368429c
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/65181
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>