LU-19066 ofd: add os_failure_domain to struct obd_statfs
Add a new field to lfs df to show the failure domain.

Update "lfs df --output=" so that when used to just
print a single field we do not pad it to a fixed width with leading
spaces.  This makes it easier to parse the outputs in tests.

Example:
  lfs df --output=domain

Also add a ltq_failure_domain field to lu_tgt_qos and set it when grab the
statfs data for an object. We do not yet use this field for anything but will
need this information later once we add failure domain awareness to the allocator.

Test-Parameters: trivial
Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>
Change-Id: Ic5bcc66b7570ad74886a550b77d91a235e72756d
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/65184
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>