LU-19157 nodemap: make nodemap name optional for banlist
Make optional the --name option to lctl nodemap_banlist_add and
nodemap_banlist_del. When omitted, the nodemap to use is inferred from
the existing NID ranges that include the provided ban range. If none
is found, the default nodemap is used.
This makes it easier for admins to ban clients, by not requiring them
to identify first in which nodemap they are.

Enhance sanity-sec test_81a to exercise this.

Test-Parameters: trivial
Test-Parameters: testlist=sanity-sec mdscount=2 mdtcount=4 osscount=1 ostcount=8 clientcount=2
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I5eefb4217105218e75e9fba4648fa30fd5f7d857
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61090
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>