LU-18033 llite: enable hybrid IO by default
Enable hybrid IO by default.  This is now passing all
tests.

Includes three small test fixes, one for sanityn 108a:
The purpose of test 108a is to confirm for async IO
(buffered writes) the write does not need to go to disk
for the file size to be updated, but for hybrid/DIO the
write does not complete until it goes to disk.

We disable hybrid here so this definitely tests buffered
IO.

One for replay-single test 89:
We need to ensure we have the dlmlock from the OST before
we take the OST offline, so add an ls -la.  Previously the
write took care of this, but hybrid IO does not use locks
when in DIO mode.

And one for sanity test 806:
ZFS does not allocate blocks immediately after a write, so
hybrid IO/direct IO cannot reliably give correct LSOM
blocks.  Disable the blocks test for ZFS and HIO.

Test-Parameters: fstype=zfs testlist=sanity
Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: Ieaed648a1a18ba36446e5ad91f6f65c150fc6330
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56855
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>