LU-19068 quota: truncated time should be kept during race
During waiting to refresh the quota usage in "qsd_upd_thread"
after some big files are deleted, the final time to wait
(lqe_truncated_time) should be kept when the lqe is added to
the queue for defer adjust.

The patch adds a fix in the test_96 to wait the deletion to be
completed to ensure the quota usage to be refreshed in time.

The patch also adds some debug output about the delayed refresh.

Test-Parameters: ostcount=2 ostsizegb=5 mdtfilesystemtype=ldiskfs ostfilesystemtype=ldiskfs testlist=sanity-quota env=ONLY=96,ONLY_REPEAT=30
Fixes: f472b887e3 ("LU-19068 quota: wait to release quota")
Signed-off-by: Hongchao Zhang <hongchao@whamcloud.com>
Change-Id: Id810327d18b910844ea71e46224d1e6563dec03e
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61162
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Sergey Cheremencev <scherementsev@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>