LU-17501 libcfs: extended range for exclude CPT pattern
The range for valid core ids was being limited to the number of cores
in each partitions, this worked for 'cpu_pattern=N C[n]' but it fails
to consider 'cpu_pattern=N X[n]' which should be able to specify any
core on the CPU.

Now the range for when 'X[n]' is used has been set to cover all
core ids on the system.

conf-sanity.sh test 200d and 200c have been modified to add checks
for this new behaviour ensuring that these options work as expected.

Test-Parameters: trivial env=ONLY="200d 200c" testlist=conf-sanity
Fixes: 6d7cf303b0 ("LU-17501 libcfs: adding X for cpu pattern")
Change-Id: Ide9f554b18c708b9f2d9fa12f0bd90933c2f1551
Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61168
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Artur Novik <anovik@ddn.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>