LU-18222 quota: add LQA to quota pool
Patch adds general commands to operate with LQA at QMT level: create,
destroy, add or remove ranges, list LQAs and ranges inside LQAs.

Each LQA is a quota pool containing a red-black tree. This red-black
tree stores quota ID ranges and is designed to verify whether an
incoming ID belongs to the current LQA. In contrast to a regular pool,
LQA has only 2 LQEs for each quota type with qids 0 and 1. The first
one is to store grace timeout. The second contains LQA limits and the
space granted by all LQA IDs. This LQE is handled together with the
global pool and pool quotas LQEs. Unlike a standard pool lookup, LQAs
are added to the list of pools(qti_pools) if contain incoming QID. At
the same time pool quotas are added to the list of pools(qti_pools) if
match the OST index of handled RPC.

Additionally, this patch fixes a bug in pool recalculation mechanism.
Even if only one of quota types is enabled, it should continue pool
recalculation. Without the fix, recalculation was done only when all
quota types are enabled(enabled=ugp).

Signed-off-by: Sergey Cheremencev <scherementsev@ddn.com>
Change-Id: I328a7e9e523ef96519404024ecd1fd88e0a7e2c5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61351
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Hongchao Zhang <hongchao@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>