LU-19242 tests: wait longer for recovery-small/26b
Add an additional wait for recovery-small test_26b to see if this
allows the test to pass, and how long it took.

This is not really a fix for the test, but it will avoid this issue
from causing problems for other paches.

Test-Parameters: trivial testlist=recovery-small env=ONLY=26b,ONLY_REPEAT=100
Fixes: 0a5ee1a2a2 ("LU-18878 ptlrpc: evictor/prolong timeout formula fix")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I93bf62e3ae25d0d28c218f8109170b28c260c731
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61288
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>