LU-19123 utils: fix 'lfs project' on bad symlink v2
Fix lfs_project_iterate() to also use lstat() instead of stat() to
handle broken symlinks.  Change the error messages for these cases to
make it more clear if someone is running the old version of this code
and is reporting a problem.

Improve sanity-quota test_64() to check this case.

Fixes: 588850bbeb ("LU-19123 utils: allow 'lfs project' on bad symlinks")
Fixes: 501e5b2c8a ("LU-18027 lfs: lfs find handling special files")
Test-Parameters: trivial testlist=sanity-quota
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Iddd9553d7e3090dbff825e17bbda8541ff9c4ae5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61309
Reviewed-by: Rahul Bansal <rahulmay94@gmail.com>
Reviewed-by: Li Dongyang <dongyangli@ddn.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>