LU-14590 obdclass: remove duplicate health_check param
Remove the duplicate /sys/kernel/debug/lustre/health_check file,
since "lctl get_param health_check" is now printing duplicate
entries for this parameter.

The debugfs "health_check" file was printing which device was
unhealthy, but only 'healthy' or 'LBUG' or 'NOT HEALTHY' is used.
The unhealthy device is printed to the console/syslog for reference,
which is probably more useful than printing it to stdout.

Fixes: 5d368bd0b2 ("LU-8066 obd: make health_check sysfs compliant")
Fixes: a2e3a2f5a3 ("LU-14590 utils: merge similar list_param params")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Iceabbc4ba3b51d1225765fb8b2ae613fc2ff8c20
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61530
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Timothy Day <timday@amazon.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Nathaniel Clark <nclark@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>