LU-19424 lmv: expose qos_rr_index parameter
Add a new lmv.*.qos_rr_index parameter that prints or sets
the current qos_rr_index value modulo the number of MDTs.

This can be used to determine if the round-robin index is
being initialized as expected, or can be set to ensure a
uniform distribution of clients across MDTs for testing.

Fixes: b601eb35e9 ("LU-11213 lmv: reuse object alloc QoS code from LOD")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I5e3176c75bbaf2b146f9bb79872a6272382540e5
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61543
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Lai Siyao <lai.siyao@whamcloud.com>
Reviewed-by: Frederick Dilger <fdilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>