LU-19443 utils: show help for invalid interactive commands
When lctl interactive mode is used, it is possible to type "help" to
print the short help for a given command. It would be helpful if
invalid commands, such as "?", also presented the short help of the
command currently being run in interactive mode.

This patch adds this functionality.

Test-Parameters: trivial
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: Ie8cb92c5e4f57f18d501dac4525f36788a03f38d
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61642
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>