LU-18939 llite: allow readahead params to be set as percentage
Allow the following readahead parameters to be set as a
percentage of total RAM, similar to max_cached_mb:
- llite.*.max_read_ahead_mb
- llite.*.max_read_ahead_per_file_mb
- llite.*.max_read_ahead_whole_mb

This allows administrators to set these parameters as a
percentage of the total RAM on the system, which is more
convenient than calculating the absolute values, especially
on systems with different amounts of RAM.

Added a test to verify the percentage-based settings work
correctly. The test also checks and temporarily increases
max_cached_mb to 75% of RAM. It verifies that values exceeding
the limits (50% of RAM for max_read_ahead_mb, max_read_ahead_mb
for max_read_ahead_per_file_mb, and max_read_ahead_per_file_mb
for max_read_ahead_whole_mb) are properly capped.

Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: I480c25af589df29a6c1e8a9504e09eb7d6c577ee
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58907
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Chakshu Kansal <ckansal@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>