LU-19453 lov: use-after-free of lov_stripe_md
With the conversion to kref() a LASSERT() that prevented a UAF
was removed. However some callers of lsm_addref() handled this
case gracefully so the LASSERT() was not in the correct location.

Remove the UAF race with kref_get_unless_zero() and returning NULL.
Callers which cannot gracefully handle the UAF path should LASSERT()

HPE-bug-id: LUS-13080
Fixes: c3fbc7d721 ("LU-16796 lov: Change struct lov_stripe_md to use kref")
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I8e301f0238140fcda33f325e8bc1653f304cc556
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61710
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>