LU-19465 ec: add erasure code as a separate library
It contains the plain-C version of the upstream library taken at version
v2.31.0-104-ge677f66
The files are mostly un-modified to make it easy to verify and compare
them to upstream, or just dropping in new/updated versions when
required.

Split the kernel module into a wrapper, ec_kmod.c, and the actual
erasure code, ec_base.c.

Test-Parameters: trivial
Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>
Change-Id: Ife875ec8c5c3a9d66f34a2cb96f452d84ce9c1ce
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62087
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Patrick Farrell <pfarrell@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@dilger.ca>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>