LU-8229 ssk: support adding multiple similar server keys
Support adding multiple server keys for the same file system and same
nodemap. This can be useful for implementing a key rotation mechanism.

Add a new '-s|--suffix' option to lgss_sk, only accepted when loading
a key, and that works on server keys only. When this option is
specified, and lgss_sk detects that a key with same descriptor (same
fs, same nodemap) is already loaded, then it would avoid replacing the
existing key, which is the default behavior, but instead insert the
new key with a timestamp appended in the form ":yyyymmdd_HHMMSS_USEC".
The former key is set to expire 2 days in the future, or after the
timeout in seconds given via the new '-x|--timeout' option to lgss_sk.
This new option is only accepted when loading a key with the
'-s|--suffix' option.

Test-Parameters: trivial
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: If097ee72a9ea6800de568ef3fe849f25e589ccb0
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62591
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>