LU-14554 lfs: Add --force-resync option for mirror resync
It should be possible to force the resync of a mirror with:
    "lfs mirror resync --only N --force-resync <file>"
in order to resync an inconsistent component.

This patch introduces the "--force-resync" option to force
resynchronization of the specified mirror component, even
when it is not marked stale.

Test cases are updated to verify the correct behavior of
the new option.

Test-Parameters: trivial testlist=sanity-flr

Signed-off-by: Chakshu Kansal <ckansal@ddn.com>
Change-Id: I405a744d67a9b7620ecad6b4b151e641f4f0417e
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61814
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@dilger.ca>