LU-19473 lnet: nid_addr_is_set false negatives
Numeric based NIDs can have address "0". This causes problems for
nid_addr_is_set() which can only look for a non-zero address value.
This causes two bugs:
 - If a node has multiple NIDs on one net, one of the NIDs has a 0
   address, then an attempt to delete the 0 address NID will cause
   all NIDs on the network to be removed.
 - If the source NID specified for lnetctl ping has a 0 address then
   this is not correctly passed down to the kernel (so the ping is
   processed as though no souce NID were specified).

To solve the first issue we pass a new flag to lnet_dyn_del_ni() to
indicate whether to allow deletion of the entire network in that
function. This is only needed for backward compatability with the
IOCTL interface. The netlink callers of this function always pass a
NID whose address has been "set".

To solve the second issue we modify jt_ping() to pass a NULL argument
to yaml_lnet_ping() when a source nid has not been specified by the
user.

A comment is added to nid_addr_is_set() to document the limitation of
this function.

Fixes: 277b07e175 ("LU-18540 lnet: Fix lnetctl ping timeout parsing")
Fixes: 8f8f6e2f36 ("LU-10003 lnet: use Netlink to support old and new NI APIs.")
Test-Parameters: trivial
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Id9c9f8ff93f7cdb040c9c166af4cdfcb64b2b6e8
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61822
Reviewed-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Reviewed-by: Cyril Bordage <cbordage@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>