LU-19491 utils: add nodemap_new alias for nodemap_add
For a better distinction between the action to create a nodemap and
the action to modify nodemap properties such as the NID ranges or the
offset, introduce a 'nodemap new' alias for 'nodemap add'.
The 'new' command takes the exact same syntax as 'add'. Document this
as the preferred command, and keep "nodemap add" for compatibility.

Update sanity-sec test_7, test_8, test_9, test_10, test_11, test_12,
test_13, test_14, test_15 to use the 'new' action.

Test-Parameters: trivial
Test-Parameters: testlist=sanity-sec
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I642410f79fe1c6c77330a7a1d0206cde730f99b1
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/62182
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@dilger.ca>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>