LU-19498 lnet: fail gracefully on too many NIs
When too many NIs get created, i.e. in excess of
lnet_interfaces_max, that triggers an LBUG in
lnet_ping_target_update().

Add the count of local NIs and fail with an error
instead of causing a crash when the limit is about
to be exceeded.

Test-Parameters: trivial testlist=sanity-lnet
Signed-off-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Change-Id: I03aad11d8ad097b0d1c655bca7cbc27a3a875cff
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61923
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Mikhail Pershin <mpershin@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@dilger.ca>