LU-19028 nodemap: do not squash trusted ids
When a root access is made with the admin property set, or when a
regular user access is made with the trusted property set, id should
not be squashed, even if it matches the defined squash_id on the
nodemap.

Add sanity-sec test_84 to verify trusted ids are not squashed.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: Id0dca0d23c29919781535df40a30f8f0c33aff8c
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61967
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>