LU-18502 utils: Support YAML for get_param parameter names
"lctl get_param -y" does not print a valid YAML for the parameter
names even if the values are YAML-compliant. This patch extends the
"-y" parameter to also affect the parameter names, such that the '='
separator is replaced by ': '. For example, the get_param default
output for nodemap 'id' and 'idmap' are:

nodemap.c0.id=2
nodemap.c0.idmap=[
 { idtype: uid, client_id: 1003, fs_id: 11000 },
 { idtype: gid, client_id: 1003, fs_id: 11000 }
]

With "lctl get_param -y" this becomes YAML-compliant:

nodemap.c0.id: 2
nodemap.c0.idmap: [
 { idtype: uid, client_id: 1003, fs_id: 11000 },
 { idtype: gid, client_id: 1003, fs_id: 11000 }
]

In the context of this patch, the empty line was removed when printing
multi-line parameters with "lctl get_param -H" to be consistent with
the other get_param modes. Further, the tabs when printing filesets
through "lctl get_param" were replaced with spaces since tabs are not
yaml-compliant. The corresponding "awk" changes in fileset-testing
utility functions are compatible with previous tab format.

Sanity-sec test_25c was added to test this new argument and nodemap
yaml-compliance for all present and future properties.

Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: I3e9d8c5bf4feb0eafd2eac7b5d5e526dda7b096d
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58113
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>